started multithreading static collision started implementing shared approach for the different static colliders
added power of two constraint

+++ include/magique/core/Types.h
@@ -165,7 +165,7 @@ namespace magique
 
     struct ColliderInfo final
     {
-        int data;
+        uint16_t data;
         ColliderType type;
     };
